fixup: cpu: x64: amx: bwd/wu convolutions: store diff_wei in vnni format
Fix compiler warnings about comparison between signed and unsigned integers.

+++ src/cpu/x64/jit_avx512_core_amx_conv_kernel.cpp
@@ -4491,7 +4491,7 @@ void jit_diff_wei_trans_to_vnni_t::generate() {
     align(64);
     L(zero_buffer);
     const uint16_t zero = 0;
-    for (size_t i = 0; i < typesize_acc * jcp.oc_block * jcp.ic_block; ++i)
+    for (int i = 0; i < typesize_acc * jcp.oc_block * jcp.ic_block; ++i)
         db(zero);
 }
